b1873: Make various operations more robust to repeats  + support IMDSv2

+++ configuration/environments_scripts/reverse_proxy/files/root/setupHttpdGitLocal.sh
@@ -1,11 +1,15 @@
 #!/bin/bash
 REMOTE=$1
 cd /etc/httpd
-rm -rf .git
-rm -rf conf   # Perhaps, this script should instead mv to a backup location
-rm -rf conf.d
-git init
-git remote add origin "${REMOTE}"
-GIT_SSH_COMMAND="ssh -A -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no"  git fetch 
-git checkout main
+git status
+if [[ "$?" -ne 0 ]]; then
+    rm -rf .git
+    rm -rf conf   # Perhaps, this script should instead mv to a backup location
+    rm -rf conf.d
+    rm -rf .gitignore
+    git init
+    git remote add origin "${REMOTE}"
+    GIT_SSH_COMMAND="ssh -A -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no"  git fetch 
+    git checkout main
+fi
